requirements.txt
fastapi~=0.115.6
uvicorn
pydantic~=2.10.3
tortoise-orm
SQLAlchemy~=2.0.36
pytest~=8.3.4